+++ simwise/navigation/attitude_determination.py
